# specs/openapi.yaml
openapi: 3.1.0
info:
  title: FASTag Chain Reaction API
  version: 0.1.0
  summary: Backend API for the FASTag Chain Reaction mobile app and NHAI toll-health dashboard.
  description: |
    The FASTag Chain Reaction service powers a driver-facing mobile app and a regulator-facing
    dashboard for India's electronic toll-collection ecosystem.

    The API covers four broad capability areas:

      * **Authentication** via phone-number one-time-passwords (OTPs).
      * **Vehicle and FASTag account management** for the logged-in user.
      * **Real-time location pings** which the backend correlates against toll geofences
        to surface "approaching toll" events, low-balance warnings, and pre-toll recharge prompts.
      * **Recharge orchestration** through UPI Intent flows plus a webhook callback from NPCI,
        and post-toll telemetry collection that feeds the aggregated toll-health dashboard
        consumed by NHAI staff.

    All authenticated endpoints expect a Bearer token issued by `POST /v1/auth/otp-verify`.
    Rate limiting is applied per-token (or per-IP for the unauthenticated OTP endpoints) and is
    surfaced via standard `X-RateLimit-*` response headers documented under each endpoint.
  contact:
    name: FASTag Chain Reaction Engineering
    email: engineering@fastag-chain-reaction.example
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.fastag-chain-reaction.example/v1
...
security:
  - BearerAuth: []
paths:
...
  /fastag-status/{vehicle_reg}:
    get:
      tags: [FASTag]
      summary: Get the current FASTag status for a vehicle
      description: |
        Returns the latest known FASTag account state for the supplied vehicle registration
        number. The data is cached for up to 60 seconds; clients can force a fresh fetch by
        passing `?fresh=true` though this is rate-limited more aggressively.
      operationId: getFastagStatus
      parameters:
        - name: vehicle_reg
          in: path
          required: true
          description: Vehicle registration number, e.g. `KA01AB1234`. Case-insensitive.
          schema:
            type: string
            example: "KA01AB1234"
        - name: fresh
          in: query
          required: false
          description: If true, bypass the read-through cache and hit the issuing bank directly.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Current FASTag status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FastagStatus'
              example:
                vehicle_registration: "KA01AB1234"
                fastag_serial: "608216-0011223344"
                issuing_bank: "ICICI"
                balance_inr: "84.50"
                status: "active"
                blacklisted: false
                last_synced_at: "2026-05-24T08:41:50Z"
                low_balance_threshold_inr: "150.00"
                low_balance: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/ServerError'
  /recharge:
    post:
      tags: [Recharge]
      summary: Initiate a FASTag recharge via UPI Intent
      description: |
        Creates a pending recharge transaction and returns a UPI Intent deep-link that the
        client opens to hand off to the user's UPI app (GPay, PhonePe, Paytm, etc). Settlement
        is asynchronous; NPCI calls `POST /v1/recharge/webhook` once the transaction reaches a
        terminal state.
      operationId: initiateRecharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RechargeRequest'
            example:
              vehicle_id: "01HV5M0P3R6Z8M2C4V6X8K1Q4N"
              amount_inr: "500.00"
              upi_handle_hint: "priya@okhdfc"
      responses:
        '201':
          description: Recharge initiated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RechargeTxn'
              example:
                id: "01HV5M3K7E9V4P2N8R5T2C0W1Q"
                vehicle_id: "01HV5M0P3R6Z8M2C4V6X8K1Q4N"
                fastag_serial: "608216-0011223344"
                amount_inr: "500.00"
                status: "initiated"
                upi_intent_url: "upi://pay?pa=fastag@npci&pn=FASTagRecharge&tr=01HV5M3K7E9V4P2N8R5T2C0W1Q&am=500.00&cu=INR"
                provider_reference: null
                created_at: "2026-05-24T08:43:02Z"
                updated_at: "2026-05-24T08:43:02Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Recharge could not be initiated by the upstream provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "provider_declined"
                  message: "Issuing bank temporarily unavailable. Try again in a minute."
        '5XX':
          $ref: '#/components/responses/ServerError'
...
  /telemetry:
    post:
      tags: [Telemetry]
      summary: Submit post-toll experience telemetry
      description: |
        Submitted by the client a few seconds after the user passes through a toll plaza.
        Captures whether the lane camera read the FASTag successfully, queue length, dwell
        time, and an optional user rating. Feeds the aggregated toll-health dashboard.
      operationId: postTelemetry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryEvent'
            example:
              vehicle_id: "01HV5M0P3R6Z8M2C4V6X8K1Q4N"
              toll_plaza_id: "TP-NH48-ELC-014"
              outcome: "read_success"
              dwell_seconds: 12
              queue_length_vehicles: 4
              rating: 4
              comment: "Smooth pass-through, no honking today."
              occurred_at: "2026-05-24T08:46:55Z"
      responses:
        '202':
          description: Telemetry accepted for aggregation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                  event_id:
                    type: string
                example:
                  accepted: true
                  event_id: "01HV5M5Q4D8B6N9V2X8K3Q1R7T"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '5XX':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token issued by `POST /v1/auth/otp-verify`. Tokens are JWTs signed with HS256
        and have a 30-day default lifetime.
    NpciWebhookSignature:
      type: apiKey
      in: header
      name: X-NPCI-Signature
      description: HMAC-SHA256 signature over the raw request body, hex-encoded.
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "validation_failed"
              message: "Field `phone` must be an E.164 formatted Indian mobile number."
              details:
                - field: "phone"
                  issue: "format"
    Unauthorized:
      description: Missing, expired, or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthorized"
              message: "Authentication required."
...
    NotFound:
      description: The requested resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "not_found"
              message: "Resource not found."
...
    ServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "internal_error"
              message: "Something went wrong on our end. The incident has been logged."
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              example: "validation_failed"
            message:
              type: string
              description: Human-readable message safe to surface to end users.
              example: "Field `phone` must be an E.164 formatted Indian mobile number."
            details:
              type: array
              description: Optional structured details, primarily for validation errors.
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "phone"
                  issue:
                    type: string
                    example: "format"
...
    FastagStatus:
      type: object
      required: [vehicle_registration, status, blacklisted, last_synced_at]
      properties:
        vehicle_registration:
          type: string
          example: "KA01AB1234"
        fastag_serial:
          type: [string, "null"]
          example: "608216-0011223344"
        issuing_bank:
          type: [string, "null"]
          example: "ICICI"
        balance_inr:
          type: [string, "null"]
          description: Current FASTag wallet balance in INR. Encoded as a fixed-point decimal string.
          example: "84.50"
        status:
          type: string
          enum: [active, low_balance, blacklisted, inactive, unknown]
          example: "active"
        blacklisted:
          type: boolean
          example: false
        last_synced_at:
          type: string
          format: date-time
          example: "2026-05-24T08:41:50Z"
        low_balance_threshold_inr:
          type: [string, "null"]
          example: "150.00"
        low_balance:
          type: boolean
          description: True if `balance_inr < low_balance_threshold_inr`.
          example: true
    RechargeRequest:
      type: object
      required: [vehicle_id, amount_inr]
      properties:
        vehicle_id:
          type: string
          example: "01HV5M0P3R6Z8M2C4V6X8K1Q4N"
        amount_inr:
          type: string
          description: Recharge amount in INR. Fixed-point decimal string, max two decimal places.
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
          example: "500.00"
        upi_handle_hint:
          type: [string, "null"]
          description: Optional UPI VPA hint to pre-populate in the UPI Intent.
          example: "priya@okhdfc"
    RechargeTxn:
      type: object
      required: [id, vehicle_id, amount_inr, status, created_at]
      properties:
        id:
          type: string
          example: "01HV5M3K7E9V4P2N8R5T2C0W1Q"
        vehicle_id:
          type: string
          example: "01HV5M0P3R6Z8M2C4V6X8K1Q4N"
        fastag_serial:
          type: [string, "null"]
          example: "608216-0011223344"
        amount_inr:
          type: string
          example: "500.00"
        status:
          type: string
          enum: [initiated, pending, success, failed, refunded]
          example: "initiated"
        upi_intent_url:
          type: [string, "null"]
          description: UPI deep-link the client should open via Android Intent / iOS Universal Link.
          example: "upi://pay?pa=fastag@npci&pn=FASTagRecharge&tr=01HV5M3K7E9V4P2N8R5T2C0W1Q&am=500.00&cu=INR"
        provider_reference:
          type: [string, "null"]
          description: NPCI / bank reference number once the transaction settles.
          example: "NPCI-20260524-87412"
        created_at:
          type: string
          format: date-time
          example: "2026-05-24T08:43:02Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-24T08:43:02Z"
...
    TelemetryEvent:
      type: object
      required: [vehicle_id, toll_plaza_id, outcome, occurred_at]
      properties:
        vehicle_id:
          type: string
          example: "01HV5M0P3R6Z8M2C4V6X8K1Q4N"
        toll_plaza_id:
          type: string
          example: "TP-NH48-ELC-014"
        outcome:
          type: string
          enum: [read_success, read_failure, manual_override, lane_redirected, low_balance_rejected]
          example: "read_success"
        dwell_seconds:
          type: [integer, "null"]
          minimum: 0
          example: 12
        queue_length_vehicles:
          type: [integer, "null"]
          minimum: 0
          example: 4
        rating:
          type: [integer, "null"]
          description: 1-5 star user rating for the toll experience.
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: [string, "null"]
          maxLength: 500
          example: "Smooth pass-through, no honking today."
        occurred_at:
          type: string
          format: date-time
          example: "2026-05-24T08:46:55Z"
...